+++ backend/antibiootit/src/main/java/fi/tuni/koodimankelit/antibiootit/services/AntibioticsService.java
@@ -3,8 +3,8 @@ package fi.tuni.koodimankelit.antibiootit.services;
 import java.util.List;
 
 import fi.tuni.koodimankelit.antibiootit.database.data.DiagnoseInfo;
-import fi.tuni.koodimankelit.antibiootit.models.Parameters;
 import fi.tuni.koodimankelit.antibiootit.models.Treatments;
+import fi.tuni.koodimankelit.antibiootit.models.request.Parameters;
 
 
 public interface AntibioticsService {
